% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avEndpoints.R
\name{checkRequest}
\alias{checkRequest}
\title{Check Get/Post calls for success messages.}
\usage{
checkRequest(request)
}
\arguments{
\item{request}{the returned object from a get or post call}
}
\value{
either null or error message
}
\description{
{Check post calls for success messages and print the error otherwise.}
}
