% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avEndpoints.R
\name{checkResponse}
\alias{checkResponse}
\title{Check Get/Post calls for success messages.}
\usage{
checkResponse(response)
}
\arguments{
\item{response}{the returned object from a json-parsed get or post call}
}
\value{
either null or error message
}
\description{
{Check post calls for AV messages and print the error otherwise.}
}
