% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssets.R
\name{getAssets}
\alias{getAssets}
\title{Download asset data from Alpha-Vantage}
\usage{
getAssets(
  symbols,
  period = "daily",
  interval = NULL,
  market = "USD",
  crypto = FALSE,
  datatype = "json",
  key = "premium"
)
}
\arguments{
\item{symbols}{vector of stock ticker or coin symbols to download}

\item{period}{period: "intraday", "daily", "weekly", "monthly"}

\item{interval}{NULL for anything but intraday otherwise "1min", "5min", "15min", "30min", "60min"}

\item{market}{only needed for crypto: use "USD", "CNY", "EUR", etc}

\item{crypto}{boolean for downloading crypto}

\item{datatype}{"json" or "csv"}

\item{key}{"premium" or "free"; the type of API key obtained from AV}
}
\value{
xts object
}
\description{
{A wrapper to Alpha Vantage's API. This function can download either stocks or crypto-currencies. Adjusted close prices
are returned when pulling multiple stocks in a single xts object with columns for each stock. Otherwise, for a single asset
the entire OHLCV object is returned.}
}
\details{
{This is a wrapper to non-exported functions that separately pull either stock or coin data.}
}
