% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoin.R
\name{getCoin}
\alias{getCoin}
\title{Download daily time series of cryptocurrency from Alpha-Vantage}
\usage{
getCoin(symbol = "BTC", market = "USD", datatype = "json", key = "premium")
}
\arguments{
\item{symbol}{cryptocurrency symbol, i.e. "BTC" (string)}

\item{market}{physical currency symbol for quoting i.e. "USD", "CNY", etc.}

\item{datatype}{"json" or "csv" for return type to be parsed}

\item{key}{string for the API key to use Alpha Vantage}
}
\value{
xts object
}
\description{
{Download cryptocurrency daily time series from Alpha vantage using
either a free or premium API key. Includes open high low close, volume and market cap.}
}
