% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoins.R
\name{getCoins}
\alias{getCoins}
\title{Load multiple cryptocurrencies}
\usage{
getCoins(symbols, period = "daily", interval = NULL, key = "premium")
}
\arguments{
\item{symbols}{list of cryptocurrency symbols to look up}

\item{period}{"daily", "weekly", "monthly", "yearly", or "intraday"}

\item{interval}{NULL for anything but intraday otherwise "1min", "5min", "15min", "30min", "60min"}

\item{key}{"premium" or "free"; the type of API key obtained from AV}
}
\value{
xts
}
\description{
{Returns common sample of historical prices of given portfolio of cryptocurrencies.}
}
