% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStocks.R
\name{getStocks}
\alias{getStocks}
\title{Load multiple stocks into a single time-series/matrix}
\usage{
getStocks(symbols, period = "daily", key = "premium")
}
\arguments{
\item{symbols}{list of stock symbols to look up}

\item{period}{resolution between prices: "daily", "weekly", "monthly"}

\item{key}{"premium" or "free"; the type of API key obtained from AV}
}
\value{
xts
}
\description{
{Returns common sample of historical prices of given portfolio.}
}
